snapshot-sidecar: Parse snapshot dir from relayer config
