+++ config/src/lib.rs
@@ -1,643 +1,8 @@
 //! Groups configurations used throughout the relayer passed to the CLI
 
+mod cli;
+pub mod parsing;
 mod token_remaps;
+mod validation;
 
-use arbitrum_client::constants::Chain;
-use circuit_types::{elgamal::DecryptionKey, fixed_point::FixedPoint};
-use clap::Parser;
-use colored::*;
-use common::types::{
-    exchange::Exchange,
-    gossip::{ClusterId, WrappedPeerId},
-    wallet::WalletIdentifier,
-};
-use ed25519_dalek::{Digest, Keypair as DalekKeypair, Sha512, SignatureError};
-use ethers::{core::rand::thread_rng, signers::LocalWallet};
-use libp2p::{identity::Keypair, Multiaddr, PeerId};
-use rand_core::OsRng;
-use serde::{Deserialize, Serialize};
-use std::{
-    collections::HashSet,
-    env, fs,
-    net::{IpAddr, SocketAddr},
-    str::FromStr,
-};
-pub use token_remaps::setup_token_remaps;
-use toml::{value::Map, Value};
-use url::Url;
-use util::arbitrum::{parse_addr_from_deployments_file, DARKPOOL_PROXY_CONTRACT_KEY};
-
-/// The dummy message used for checking elliptic curve key pairs
-const DUMMY_MESSAGE: &str = "signature check";
-/// The CLI argument name for the config file
-const CONFIG_FILE_ARG: &str = "--config-file";
-
-// -------
-// | CLI |
-// -------
-
-/// Defines the relayer system command line interface
-#[derive(Debug, Parser, Serialize, Deserialize)]
-#[clap(author, about, long_about = None)]
-#[rustfmt::skip]
-struct Cli {
-    // ---------------
-    // | Config File |
-    // ---------------
-    /// An auth config file to read from
-    #[clap(long, value_parser)]
-    pub config_file: Option<String>,
-    /// The price reporter from which to stream prices.
-    /// If unset, the relayer will connect to exchanges directly.
-    #[clap(long, value_parser, conflicts_with_all = &["coinbase_api_key", "coinbase_api_secret", "eth_websocket_addr"])]
-    pub price_reporter_url: Option<String>,
-
-    // -----------------------------
-    // | Application Level Configs |
-    // -----------------------------
-
-    /// The take rate of this relayer on a managed match, i.e. the amount of the received asset 
-    /// that the relayer takes as a fee
-    /// 
-    /// Defaults to 8 basis points
-    #[clap(long, value_parser, default_value = "0.0008")]
-    pub match_take_rate: f64,
-    /// The mutual exclusion list for matches, two wallets in this list will never be matched internally by the node
-    #[clap(long, value_parser, value_delimiter=' ', num_args=0..)]
-    pub match_mutual_exclusion_list: Vec<WalletIdentifier>,
-
-    // -----------------------
-    // | Environment Configs |
-    // -----------------------
-
-    /// The chain that the relayer settles to
-    #[clap(long, value_parser, default_value = "testnet")]
-    pub chain_id: Chain,
-    /// The address of the darkpool contract, defaults to the internal testnet deployment
-    #[clap(long, value_parser, default_value = "0x84bc75bc4fb943d5576545fb4724c5cab781a61e")]
-    pub contract_address: String,
-    /// The path to the file containing deployments info for the darkpool contract
-    #[clap(long, value_parser)]
-    pub deployments_file: Option<String>,
-    /// The path to the file containing token remaps for the given chain
-    /// 
-    /// See https://github.com/renegade-fi/token-mappings for more information on the format of this file
-    #[clap(long, value_parser)]
-    pub token_remap_file: Option<String>,
-
-    // ----------------------------
-    // | Networking Configuration |
-    // ----------------------------
-
-    /// Allow for discovery of nodes on the localhost IP address
-    #[clap(long, value_parser, default_value="false")]
-    pub allow_local: bool,
-    /// The address to bind to for gossip, defaults to 0.0.0.0 (all interfaces)
-    #[clap(long, value_parser, default_value = "0.0.0.0")]
-    pub bind_addr: IpAddr,
-    /// The known public IP address of the local peer
-    #[clap(long, value_parser)] 
-    pub public_ip: Option<SocketAddr>,
-    /// The amount of time to allow for gossip warmup, in milliseconds
-    /// 
-    /// Defaults to 10s
-    #[clap(long, value_parser, default_value = "10000")]
-    pub gossip_warmup: u64,
-    
-    // -------------------------
-    // | Cluster Configuration |
-    // -------------------------
-    /// The bootstrap servers that the peer should dial initially
-    #[clap(short, long, value_parser)]
-    pub bootstrap_servers: Option<Vec<String>>,
-    /// The cluster private key to use
-    #[clap(long = "cluster-private-key", value_parser)]
-    pub cluster_private_key: Option<String>,
-    /// The cluster public key to use
-    #[clap(long = "cluster-public-key", value_parser)]
-    pub cluster_public_key: Option<String>,
-
-    // ----------------------------
-    // | Local Node Configuration |
-    // ----------------------------
-
-    /// The port to listen on for libp2p
-    #[clap(short = 'p', long, value_parser, default_value = "8000")]
-    pub p2p_port: u16,
-    /// The port to listen on for the externally facing HTTP API
-    #[clap(long, value_parser, default_value = "3000")]
-    pub http_port: u16,
-    /// The port to listen on for the externally facing websocket API
-    #[clap(long, value_parser, default_value = "4000")]
-    pub websocket_port: u16,
-    /// The local peer's base64 encoded p2p key
-    /// A fresh key is generated at startup if this is not present
-    #[clap(long, value_parser)]
-    pub p2p_key: Option<String>,
-    /// The path at which to open up the database
-    #[clap(long, value_parser, default_value = "./relayer_state.db")]
-    pub db_path: String,
-    /// The path at which to save raft snapshots
-    #[clap(long, value_parser, default_value = "./raft_snapshots")]
-    pub raft_snapshot_path: String,
-    /// The maximum staleness (number of newer roots observed) to allow on Merkle proofs for 
-    /// managed wallets. After this threshold is exceeded, the Merkle proof will be updated
-    #[clap(long, value_parser, default_value = "100")]
-    pub max_merkle_staleness: usize,
-    /// Flag to disable the price reporter
-    #[clap(long, value_parser)]
-    pub disable_price_reporter: bool,
-    /// Disables exchanges for price reporting
-    #[clap(long, value_parser, num_args=1.., value_delimiter=' ')]
-    pub disabled_exchanges: Vec<Exchange>,
-    /// Whether or not to run the relayer in debug mode
-    #[clap(short, long, value_parser)]
-    pub debug: bool,
-
-    // -----------
-    // | Secrets |
-    // -----------
-    /// The Coinbase API key to use for price streaming
-    #[clap(long = "coinbase-key", value_parser)]
-    pub coinbase_api_key: Option<String>,
-    /// The Coinbase API secret to use for price streaming
-    #[clap(long = "coinbase-secret", value_parser)]
-    pub coinbase_api_secret: Option<String>,
-    /// The Ethereum RPC node websocket address to dial for on-chain data
-    #[clap(long = "eth-websocket", value_parser)]
-    pub eth_websocket_addr: Option<String>,
-    /// The HTTP addressable Arbitrum JSON-RPC node
-    #[clap(long = "rpc-url", value_parser)]
-    pub rpc_url: Option<String>,
-    /// The Arbitrum private keys used to send transactions.
-    /// Multiple keys can be provided to mitigate nonce contention across a node / cluster.
-    /// 
-    /// Defaults to the devnet pre-funded key
-    #[clap(long = "arbitrum-pkeys", value_parser, default_values_t = ["0xb6b15c8cb491557369f3c7d2c287b053eb229daa9c22138887752191c9520659".to_string()])]
-    pub arbitrum_private_keys: Vec<String>,
-    /// The key used to decrypt fee payments
-    #[clap(long = "fee-decryption-key", value_parser)]
-    pub fee_decryption_key: Option<String>,
-
-    // -------------
-    // | Telemetry |
-    // -------------
-
-    /// Whether or not to enable OTLP tracing
-    #[clap(long = "enable-otlp", value_parser)]
-    pub otlp_enabled: bool,
-    /// The OTLP collector endpoint to send traces to
-    #[clap(long, value_parser, default_value = "http://localhost:4317")]
-    pub otlp_collector_url: String,
-    /// Whether or not to enable Datadog-formatted logs
-    #[clap(long = "enable-datadog", value_parser)]
-    pub datadog_enabled: bool,
-    /// Whether or not to enable metrics collection
-    #[clap(long = "enable-metrics", value_parser)]
-    pub metrics_enabled: bool,
-    /// The StatsD recorder host to send metrics to
-    #[clap(long, value_parser, default_value = "127.0.0.1")]
-    pub statsd_host: String,
-
-    /// The StatsD recorder port to send metrics to
-    #[clap(long, value_parser, default_value = "8125")]
-    pub statsd_port: u16,
-}
-
-// ----------
-// | Config |
-// ----------
-
-/// Defines the system config for the relayer
-#[derive(Debug)]
-pub struct RelayerConfig {
-    // -----------------------------
-    // | Application Level Configs |
-    // -----------------------------
-    /// The take rate of this relayer on a managed match, i.e. the amount of the
-    /// received asset that the relayer takes as a fee
-    pub match_take_rate: FixedPoint,
-    /// The mutual exclusion list for matches, two wallets in this list will
-    /// never be matched internally by the node
-    pub match_mutual_exclusion_list: HashSet<WalletIdentifier>,
-    /// The price reporter from which to stream prices.
-    /// If unset, the relayer will connect to exchanges directly.
-    pub price_reporter_url: Option<Url>,
-
-    // -----------------------
-    // | Environment Configs |
-    // -----------------------
-    /// The chain that the relayer settles to
-    pub chain_id: Chain,
-    /// The address of the contract in the target network
-    pub contract_address: String,
-
-    // ----------------------------
-    // | Networking Configuration |
-    // ----------------------------
-    /// Allow for discovery of nodes on the localhost IP address
-    pub allow_local: bool,
-    /// The address to bind to for gossip, defaults to 0.0.0.0 (all interfaces)
-    pub bind_addr: IpAddr,
-    /// The known public IP address of the local peer
-    pub public_ip: Option<SocketAddr>,
-    /// The amount of time to allow for gossip warmup, in milliseconds
-    pub gossip_warmup: u64,
-
-    // -------------------------
-    // | Cluster Configuration |
-    // -------------------------
-    /// Bootstrap servers that the peer should connect to
-    pub bootstrap_servers: Vec<(WrappedPeerId, Multiaddr)>,
-    /// The cluster keypair
-    pub cluster_keypair: DalekKeypair,
-
-    // ----------------------------
-    // | Local Node Configuration |
-    // ----------------------------
-    /// The port to listen on for libp2p
-    pub p2p_port: u16,
-    /// The port to listen on for the externally facing HTTP API
-    pub http_port: u16,
-    /// The port to listen on for the externally facing websocket API
-    pub websocket_port: u16,
-    /// The local peer's base64 encoded p2p key
-    pub p2p_key: Keypair,
-    /// The path at which to open up the database
-    pub db_path: String,
-    /// The path at which to save raft snapshots
-    pub raft_snapshot_path: String,
-    /// The maximum staleness (number of newer roots observed) to allow on
-    /// Merkle proofs for managed wallets. After this threshold is exceeded,
-    /// the Merkle proof will be updated
-    pub max_merkle_staleness: usize,
-    /// Whether to disable the price reporter if e.g. we are streaming from a
-    /// dedicated external API gateway node in the cluster
-    pub disable_price_reporter: bool,
-    /// The exchanges explicitly disabled for price reports
-    pub disabled_exchanges: Vec<Exchange>,
-    /// Whether or not the relayer is in debug mode
-    pub debug: bool,
-
-    // -----------
-    // | Secrets |
-    // -----------
-    /// The cluster ID, a parsed version of the cluster's pubkey
-    pub cluster_id: ClusterId,
-    /// The Coinbase API key to use for price streaming
-    pub coinbase_api_key: Option<String>,
-    /// The Coinbase API secret to use for price streaming
-    pub coinbase_api_secret: Option<String>,
-    /// The HTTP addressable Arbitrum JSON-RPC node
-    pub rpc_url: Option<String>,
-    /// The Arbitrum private keys used to send transactions
-    pub arbitrum_private_keys: Vec<LocalWallet>,
-    /// The Ethereum RPC node websocket address to dial for on-chain data
-    pub eth_websocket_addr: Option<String>,
-    /// The decryption key used to settle managed match fees
-    pub fee_decryption_key: DecryptionKey,
-
-    // -------------
-    // | Telemetry |
-    // -------------
-    /// Whether or not to enable OTLP tracing
-    pub otlp_enabled: bool,
-    /// The OTLP collector endpoint to send traces to
-    pub otlp_collector_url: String,
-    /// Whether or not to enable Datadog-formatted logs
-    pub datadog_enabled: bool,
-    /// Whether or not to enable metrics collection
-    pub metrics_enabled: bool,
-    /// The StatsD recorder host to send metrics to
-    pub statsd_host: String,
-    /// The StatsD recorder port to send metrics to
-    pub statsd_port: u16,
-}
-
-impl RelayerConfig {
-    /// Get the peer ID associated with the p2p key
-    pub fn peer_id(&self) -> WrappedPeerId {
-        WrappedPeerId(self.p2p_key.public().to_peer_id())
-    }
-
-    /// Get the Arbitrum private key from which the relayer's wallet is derived
-    pub fn relayer_arbitrum_key(&self) -> &LocalWallet {
-        self.arbitrum_private_keys.first().expect("no arbitrum private keys configured")
-    }
-}
-
-impl Default for RelayerConfig {
-    fn default() -> Self {
-        // Parse a dummy set of command line args and convert this to a config
-        let cli = Cli::parse_from(Vec::<String>::new());
-        parse_config_from_args(cli).expect("default config does not parse")
-    }
-}
-
-/// A custom clone implementation specifically for the cluster keypair which
-/// does not implement clone
-impl Clone for RelayerConfig {
-    fn clone(&self) -> Self {
-        Self {
-            match_take_rate: self.match_take_rate,
-            match_mutual_exclusion_list: self.match_mutual_exclusion_list.clone(),
-            price_reporter_url: self.price_reporter_url.clone(),
-            chain_id: self.chain_id,
-            contract_address: self.contract_address.clone(),
-            bootstrap_servers: self.bootstrap_servers.clone(),
-            p2p_port: self.p2p_port,
-            http_port: self.http_port,
-            websocket_port: self.websocket_port,
-            p2p_key: self.p2p_key.clone(),
-            db_path: self.db_path.clone(),
-            raft_snapshot_path: self.raft_snapshot_path.clone(),
-            max_merkle_staleness: self.max_merkle_staleness,
-            allow_local: self.allow_local,
-            bind_addr: self.bind_addr,
-            public_ip: self.public_ip,
-            gossip_warmup: self.gossip_warmup,
-            disable_price_reporter: self.disable_price_reporter,
-            disabled_exchanges: self.disabled_exchanges.clone(),
-            cluster_keypair: DalekKeypair::from_bytes(&self.cluster_keypair.to_bytes()).unwrap(),
-            cluster_id: self.cluster_id.clone(),
-            coinbase_api_key: self.coinbase_api_key.clone(),
-            coinbase_api_secret: self.coinbase_api_secret.clone(),
-            rpc_url: self.rpc_url.clone(),
-            arbitrum_private_keys: self.arbitrum_private_keys.clone(),
-            fee_decryption_key: self.fee_decryption_key,
-            eth_websocket_addr: self.eth_websocket_addr.clone(),
-            debug: self.debug,
-            otlp_enabled: self.otlp_enabled,
-            otlp_collector_url: self.otlp_collector_url.clone(),
-            datadog_enabled: self.datadog_enabled,
-            metrics_enabled: self.metrics_enabled,
-            statsd_host: self.statsd_host.clone(),
-            statsd_port: self.statsd_port,
-        }
-    }
-}
-
-/// Parses command line args into the node config
-///
-/// We allow for configurations to come from both a config file and overrides
-/// on the command line directly. To support this, we first read configuration
-/// options from the config file, prepend them to the cli args string, and parse
-/// using the `overrides_with("self")` option so that cli args (which come after
-/// config file args) take precedence.
-pub fn parse_command_line_args() -> Result<RelayerConfig, String> {
-    // Parse args from command line and config file, place the config file args
-    // *before* the command line args so that clap will give precedence to the
-    // command line arguments
-    // However, the first argument from the command line is the executable name, so
-    // place this before all args
-    let mut command_line_args: Vec<String> =
-        env::args_os().map(|val| val.to_str().unwrap().to_string()).collect();
-    let config_file_args = config_file_args(&command_line_args)?;
-
-    let mut full_args = vec![command_line_args.remove(0)];
-    full_args.extend(config_file_args);
-    full_args.extend(command_line_args);
-
-    let cli = Cli::parse_from(full_args);
-    // Setup the token remap
-    setup_token_remaps(cli.token_remap_file.clone(), cli.chain_id)?;
-
-    let config = parse_config_from_args(cli)?;
-    Ok(config)
-}
-
-/// Parse the config from a set of command line arguments
-///
-/// Separating out this functionality allows us to easily inject custom args
-/// apart from what is specified on the command line
-fn parse_config_from_args(cli_args: Cli) -> Result<RelayerConfig, String> {
-    // Parse the cluster keypair from CLI args
-    // dalek library expects a packed byte array of [PRIVATE_KEY||PUBLIC_KEY]
-    let keypair = if cli_args.cluster_public_key.is_some() && cli_args.cluster_private_key.is_some()
-    {
-        let mut public_key: Vec<u8> = base64::decode(cli_args.cluster_public_key.unwrap()).unwrap();
-        let mut private_key: Vec<u8> =
-            base64::decode(cli_args.cluster_private_key.unwrap()).unwrap();
-        private_key.append(&mut public_key);
-
-        let keypair = ed25519_dalek::Keypair::from_bytes(&private_key[..]).unwrap();
-
-        // Verify that the keypair represents a valid elliptic curve pair
-        if validate_cluster_keypair(&keypair).is_err() {
-            panic!("cluster keypair invalid")
-        }
-
-        keypair
-    } else {
-        let mut rng = OsRng {};
-        DalekKeypair::generate(&mut rng)
-    };
-
-    // Parse the local relayer's arbitrum wallet from the cli
-    let arbitrum_private_keys = cli_args
-        .arbitrum_private_keys
-        .iter()
-        .map(|k| LocalWallet::from_str(k).map_err(|e| e.to_string()))
-        .collect::<Result<Vec<_>, _>>()?;
-    let fee_decryption_key = parse_decryption_key(cli_args.fee_decryption_key)?;
-
-    // Parse the p2p keypair or generate one
-    let p2p_key = if let Some(keypair) = cli_args.p2p_key {
-        let decoded = base64::decode(keypair).expect("p2p key formatted incorrectly");
-        Keypair::from_protobuf_encoding(&decoded).expect("error parsing p2p key")
-    } else {
-        Keypair::generate_ed25519()
-    };
-
-    let cluster_id = ClusterId::new(&keypair.public);
-
-    // Parse the bootstrap servers into multiaddrs
-    let mut parsed_bootstrap_addrs: Vec<(WrappedPeerId, Multiaddr)> = Vec::new();
-    for addr in cli_args.bootstrap_servers.unwrap_or_default().iter() {
-        let parsed_addr: Multiaddr =
-            addr.parse().expect("Invalid address passed as --bootstrap-server");
-        let peer_id = PeerId::try_from_multiaddr(&parsed_addr)
-            .expect("Invalid address passed as --bootstrap-server");
-        parsed_bootstrap_addrs.push((WrappedPeerId(peer_id), parsed_addr));
-    }
-
-    // Parse the price reporter URL, if there is one
-    let price_reporter_url = cli_args
-        .price_reporter_url
-        .map(|url| Url::parse(&url).expect("Invalid price reporter URL"));
-
-    let mut config = RelayerConfig {
-        match_take_rate: FixedPoint::from_f64_round_down(cli_args.match_take_rate),
-        match_mutual_exclusion_list: cli_args.match_mutual_exclusion_list.into_iter().collect(),
-        price_reporter_url,
-        chain_id: cli_args.chain_id,
-        contract_address: cli_args.contract_address,
-        bootstrap_servers: parsed_bootstrap_addrs,
-        p2p_port: cli_args.p2p_port,
-        http_port: cli_args.http_port,
-        websocket_port: cli_args.websocket_port,
-        allow_local: cli_args.allow_local,
-        max_merkle_staleness: cli_args.max_merkle_staleness,
-        p2p_key,
-        db_path: cli_args.db_path,
-        raft_snapshot_path: cli_args.raft_snapshot_path,
-        bind_addr: cli_args.bind_addr,
-        public_ip: cli_args.public_ip,
-        gossip_warmup: cli_args.gossip_warmup,
-        disable_price_reporter: cli_args.disable_price_reporter,
-        disabled_exchanges: cli_args.disabled_exchanges,
-        cluster_keypair: keypair,
-        cluster_id,
-        coinbase_api_key: cli_args.coinbase_api_key,
-        coinbase_api_secret: cli_args.coinbase_api_secret,
-        rpc_url: cli_args.rpc_url,
-        arbitrum_private_keys,
-        fee_decryption_key,
-        eth_websocket_addr: cli_args.eth_websocket_addr,
-        debug: cli_args.debug,
-        otlp_enabled: cli_args.otlp_enabled,
-        otlp_collector_url: cli_args.otlp_collector_url,
-        datadog_enabled: cli_args.datadog_enabled,
-        metrics_enabled: cli_args.metrics_enabled,
-        statsd_host: cli_args.statsd_host,
-        statsd_port: cli_args.statsd_port,
-    };
-
-    set_contract_from_file(&mut config, cli_args.deployments_file)?;
-    Ok(config)
-}
-
-/// Parse args from a config file
-fn config_file_args(cli_args: &[String]) -> Result<Vec<String>, String> {
-    // Find a match for the config file argument
-    let mut found = false;
-    let mut index = 0;
-
-    for arg in cli_args.iter() {
-        index += 1;
-        // If we find "--config-file", the next argument is the file to read from
-        if arg == CONFIG_FILE_ARG {
-            found = true;
-            break;
-        }
-    }
-
-    // No config file found
-    if !found {
-        return Ok(vec![]);
-    }
-
-    // Read in the config file
-    let file_contents =
-        fs::read_to_string(cli_args[index].clone()).map_err(|err| err.to_string())?;
-
-    let config_kv_pairs: Map<_, _> =
-        toml::from_str(&file_contents).map_err(|err| err.to_string())?;
-
-    let mut config_file_args: Vec<String> = Vec::with_capacity(config_kv_pairs.len());
-    for (toml_key, value) in config_kv_pairs.iter() {
-        // Format the TOML key into --key
-        let cli_arg = format!("--{}", toml_key);
-
-        // Parse the values for this TOML entry into a CLI-style vector of strings
-        let values: Vec<String> = match value {
-            // Just the flag, i.e. --flag
-            Value::Boolean(_) => vec![cli_arg],
-            // Parse all values into multiple repetitions, i.e. --key val1 --key val2 ...
-            Value::Array(arr) => {
-                let mut res: Vec<String> = Vec::new();
-                for val in arr.iter() {
-                    res.push(cli_arg.clone());
-                    res.push(toml_value_to_string(val)?);
-                }
-
-                res
-            },
-            // All other type may simply be parsed as --key val
-            _ => {
-                vec![
-                    cli_arg.clone(),
-                    toml_value_to_string(value).map_err(|_| {
-                        format!("error parsing config value: {:?} = {:?}", cli_arg, value)
-                    })?,
-                ]
-            },
-        };
-
-        config_file_args.extend(values);
-    }
-
-    Ok(config_file_args)
-}
-
-/// Helper method to convert a toml value to a string
-fn toml_value_to_string(val: &Value) -> Result<String, String> {
-    Ok(match val {
-        Value::String(val) => val.clone(),
-        Value::Integer(val) => format!("{:?}", val),
-        Value::Float(val) => format!("{:?}", val),
-        Value::Boolean(val) => format!("{:?}", val),
-        _ => {
-            return Err("unsupported value".to_string());
-        },
-    })
-}
-
-/// Parse the contract address from a deployments file, overriding the default
-/// value in the config
-fn set_contract_from_file(config: &mut RelayerConfig, file: Option<String>) -> Result<(), String> {
-    // Do not override if the file is not specified
-    if let Some(path) = file {
-        let darkpool_addr = parse_addr_from_deployments_file(&path, DARKPOOL_PROXY_CONTRACT_KEY)
-            .map_err(|e| e.to_string())?;
-        config.contract_address = darkpool_addr;
-    }
-
-    Ok(())
-}
-
-/// Parse the relayer's decryption key from a string
-pub fn parse_decryption_key(key_str: Option<String>) -> Result<DecryptionKey, String> {
-    if let Some(k) = key_str {
-        DecryptionKey::from_hex_str(&k)
-    } else {
-        // Must print here as logger is not yet setup
-        println!("{}\n", "WARN: No fee decryption key provided, generating one".yellow());
-        Ok(DecryptionKey::random(&mut thread_rng()))
-    }
-}
-
-/// Runtime validation of the keypair passed into the relayer via config
-/// Sign a simple request and verify the signature
-///
-/// The public interface does not allow us to more directly check the keypair
-/// as public_key == private_key * ed25519_generator, so we opt for this
-/// instead. Happens once at startup so we are not concerned with performance
-fn validate_cluster_keypair(keypair: &DalekKeypair) -> Result<(), SignatureError> {
-    // Hash the message
-    let mut hash_digest: Sha512 = Sha512::new();
-    hash_digest.update(DUMMY_MESSAGE);
-
-    // Sign and verify with keypair
-    let sig = keypair.sign_prehashed(hash_digest, None /* context */).unwrap();
-
-    // Rehash, hashes are not clonable
-    let mut second_hash: Sha512 = Sha512::new();
-    second_hash.update(DUMMY_MESSAGE);
-    keypair.verify_prehashed(second_hash, None /* context */, &sig)
-}
-
-// ---------
-// | Tests |
-// ---------
-
-#[cfg(test)]
-mod test {
-    use crate::RelayerConfig;
-
-    /// Test that the default config parses
-    #[test]
-    fn test_default_config() {
-        RelayerConfig::default();
-    }
-}
+pub use cli::*;
