+++ snapshot-sidecar/src/main.rs
@@ -8,7 +8,8 @@
 #![allow(incomplete_features)]
 
 use std::{
-    path::Path,
+    fs,
+    path::PathBuf,
     sync::mpsc::channel,
     time::{Duration, Instant},
 };
@@ -16,6 +17,7 @@ use std::{
 use aws_config::Region;
 use aws_sdk_s3::{primitives::ByteStream, Client};
 use clap::Parser;
+use config::{parsing::parse_config_from_file, RelayerConfig};
 use external_api::http::admin::{IsLeaderResponse, IS_LEADER_ROUTE};
 use notify::{event::ModifyKind, Config, EventKind, RecommendedWatcher, RecursiveMode, Watcher};
 use reqwest::Client as HttpClient;
@@ -25,12 +27,15 @@ use util::{
     telemetry::{setup_system_logger, LevelFilter},
 };
 
+/// The postfix of the snapshot path
+const SNAPSHOT_FILE_NAME: &str = "snapshot.gz";
+
 /// The sidecar CLI
 #[derive(Debug, Parser)]
 struct Cli {
-    /// The path to watch for snapshot files
-    #[clap(short, long)]
-    path: String,
+    /// The path to the relayer's config
+    #[clap(long)]
+    config_path: String,
     /// The name of the s3 bucket to send snapshots to
     #[clap(short, long)]
     bucket: String,
@@ -53,18 +58,29 @@ struct Cli {
 async fn main() {
     // Parse the CLI
     let cli = Cli::parse();
-    let path = Path::new(&cli.path);
+    let relayer_config =
+        parse_config_from_file(&cli.config_path).expect("could not parse relayer config");
     setup_system_logger(LevelFilter::INFO);
 
     let region = Region::new(cli.region.clone());
     let config = aws_config::from_env().region(region).load().await;
     let s3_client = aws_sdk_s3::Client::new(&config);
 
+    // If the watch path does not exist, create it as an empty file
+    let path = snapshot_path(&relayer_config);
+    if !path.exists() {
+        if let Some(parent) = path.parent() {
+            fs::create_dir_all(parent).expect("Failed to create parent directories");
+        }
+
+        fs::write(&path, []).expect("Failed to write to file");
+    }
+
     // Build a notification channel
     let (tx, rx) = channel();
     let mut watcher =
         RecommendedWatcher::new(tx, Config::default()).expect("Failed to create watcher");
-    watcher.watch(path, RecursiveMode::NonRecursive).expect("Failed to watch path");
+    watcher.watch(&path, RecursiveMode::NonRecursive).expect("Failed to watch path");
 
     // Listen for events
     let debounce_interval = Duration::from_secs(cli.interval);
@@ -72,7 +88,7 @@ async fn main() {
     for event in rx.iter().flatten() {
         if let EventKind::Modify(ModifyKind::Data(..)) = event.kind {
             if Instant::now() - last_event > debounce_interval {
-                maybe_record_snapshot(&cli, &s3_client).await;
+                maybe_record_snapshot(&cli, &relayer_config, &s3_client).await;
                 last_event = Instant::now();
             }
         }
@@ -80,7 +96,7 @@ async fn main() {
 }
 
 /// Check if the local node is the leader and record the snapshot if so
-async fn maybe_record_snapshot(args: &Cli, s3_client: &Client) {
+async fn maybe_record_snapshot(args: &Cli, conf: &RelayerConfig, s3_client: &Client) {
     // Check if the local relayer is the leader first
     match check_leader(&args.http_addr).await {
         Ok(false) => {
@@ -95,16 +111,22 @@ async fn maybe_record_snapshot(args: &Cli, s3_client: &Client) {
     };
 
     // Record the snapshot
-    let path = Path::new(&args.path);
-    if let Err(e) = handle_new_snapshot(path, &args.bucket, s3_client).await {
+    if let Err(e) = handle_new_snapshot(&args.bucket, conf, s3_client).await {
         error!("Failed to handle new snapshot: {e}");
     }
 }
 
 /// Copy a snapshot to s3
-async fn handle_new_snapshot(path: &Path, bucket: &str, s3_client: &Client) -> Result<(), String> {
+async fn handle_new_snapshot(
+    bucket: &str,
+    conf: &RelayerConfig,
+    s3_client: &Client,
+) -> Result<(), String> {
+    // Build the file path
     let ts = get_current_time_millis();
-    let file_name = format!("snapshot-{ts}.zip");
+    let path = snapshot_path(conf);
+    let dir = format!("cluster-{}", conf.cluster_id);
+    let file_name = format!("{dir}/snapshot-{ts}.gz");
     info!("uploading snapshot: {file_name} to {bucket}");
 
     // Send the file at `path` to the s3 bucket
@@ -129,4 +151,12 @@ async fn check_leader(api_base: &str) -> Result<bool, String> {
         .await
         .map(|r| r.leader)
         .map_err(|e| format!("Failed to parse response: {}", e))
+}
+
+/// Build the full path of the snapshot file
+fn snapshot_path(conf: &RelayerConfig) -> PathBuf {
+    let snap_path = conf.raft_snapshot_path.clone();
+    let dir = if snap_path.ends_with('/') { snap_path } else { format!("{snap_path}/") };
+    let full_path = format!("{}{}", dir, SNAPSHOT_FILE_NAME);
+    PathBuf::from(full_path)
 }
